Use variable instead of strings. Move upcast and downcast functions to utils.
